Move issue label operations to issue service package
